Support caching of credentials (#4)
Supports reusing the same credential on the new connections made for the
same host. This will reduce the time needed for authentication contexts
like Kerberos to talk to the KDC on each connection that is being set up
for the server. For example to create a new session there are at least 3
threads that will create a new connection where before Kerberos would
get a new TGT for each thread. Now the TGT is only retrieved for the
first connection and then reused for the subsequent ones.

As part of this change the authentication code is moved into its own
namespace that has no interdependencies on the global state configured
by PowerShell. This is useful for the future where the code may be
extracted out of the module into a standalone nuget package.

There are some other tweaks and changes in the same vein to make it
easier to extract the PowerShell specific bits from the WSMan code in
the future and other bugfixes.